+++ src/app/(normal)/article/[id]/page.tsx
@@ -2,12 +2,12 @@ import type { Metadata } from 'next';
 
 import { ArticleDetailPage } from '@/app/(normal)/article/[id]/pages/DetailPage.js';
 import { KeyType } from '@/constants/enum.js';
+import { createArticlePageMetadataById } from '@/helpers/createPageMetadata.js';
 import { isBotRequest } from '@/helpers/isBotRequest.js';
 import { memoizeWithRedis } from '@/helpers/memoizeWithRedis.js';
-import { getArticleOGById } from '@/services/getArticleOGById.js';
 
-const getArticleOGByIdRedis = memoizeWithRedis(getArticleOGById, {
-    key: KeyType.GetArticleOGById,
+const createMetadata = memoizeWithRedis(createArticlePageMetadataById, {
+    key: KeyType.CreateArticlePageMeatadaById,
 });
 
 interface Props {
@@ -17,7 +17,7 @@ interface Props {
 }
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
-    return getArticleOGByIdRedis(params.id);
+    return createMetadata(params.id);
 }
 
 export default function Page(props: Props) {
